+++ src/backend/marsha/core/models/video.py
@@ -19,6 +19,7 @@ from safedelete.queryset import SafeDeleteQueryset
 
 from marsha.core.defaults import (
     APPROVAL,
+    CELERY_PIPELINE,
     DELETED,
     DELETED_VIDEOS_STORAGE_BASE_DIRECTORY,
     ENDED,
@@ -29,6 +30,7 @@ from marsha.core.defaults import (
     LIVE_CHOICES,
     LIVE_TYPE_CHOICES,
     PENDING,
+    PROCESS_PIPELINE_CHOICES,
     RUNNING,
     STOPPING,
     TRANSCODE_PIPELINE_CHOICES,
@@ -692,6 +694,14 @@ class Thumbnail(AbstractImage):
         on_delete=models.CASCADE,
     )
 
+    process_pipeline = models.CharField(
+        max_length=255,
+        verbose_name=_("process pipeline"),
+        help_text=_("Process pipeline used to process the thumbnail"),
+        choices=PROCESS_PIPELINE_CHOICES,
+        default=CELERY_PIPELINE,
+    )
+
     @property
     def playlist(self):
         """Return the playlist of the video."""
@@ -743,6 +753,38 @@ class Thumbnail(AbstractImage):
         """
         stamp = stamp or to_timestamp(self.uploaded_on)
         return f"{self.video.pk}/thumbnail/{self.pk}/{stamp}"
+
+    def get_videos_storage_prefix(
+        self,
+        stamp=None,
+        base_dir: VIDEOS_STORAGE_BASE_DIRECTORY = VOD_VIDEOS_STORAGE_BASE_DIRECTORY,
+    ):
+        """Compute the videos storage prefix for the thumbnail.
+
+        Parameters
+        ----------
+        stamp: Type[string]
+            Passing a value for this argument will return the videos storage prefix for the shared
+            live media assuming its active stamp is set to this value. This is useful to create
+            an upload policy for this prospective version of the thumbnail, so that the
+            client can upload the file and celery task can set the `uploaded_on` field to this
+            value.
+
+        base_dir: Type[VIDEOS_STORAGE_BASE_DIRECTORY]
+            The videos storage base directory. Defaults to VOD. It will be used to compute the
+            videos storage prefix.
+
+        Returns
+        -------
+        string
+            The videos storage prefix for the thumbnail.
+        """
+        stamp = stamp or to_timestamp(self.uploaded_on)
+        base = base_dir
+        if base_dir == DELETED_VIDEOS_STORAGE_BASE_DIRECTORY:
+            base = f"{DELETED_VIDEOS_STORAGE_BASE_DIRECTORY}/{VOD_VIDEOS_STORAGE_BASE_DIRECTORY}"
+
+        return f"{base}/{self.video.pk}/thumbnail/{stamp}"
 
     def update_upload_state(self, upload_state, uploaded_on, **extra_parameters):
         """Manage upload state.