+++ src/frontend/packages/lib_video/src/components/common/VideoWidgetProvider/widgets/WidgetThumbnail/index.spec.tsx
@@ -151,6 +151,7 @@ describe('<DashboardLiveWidgetThumbnail />', () => {
       mockedThumbnail.id,
       file,
       mockedVideo.id,
+      expect.any(Function),
     );
   });